% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{adq_write}
\alias{adq_write}
\title{Applies the Adequacy and write study by mc year}
\usage{
adq_write(
  sim_opts,
  areas,
  virtual_areas,
  links_NTC_data,
  ptdf_FB_data,
  capacity_FB_data,
  ts_FB_data,
  mcYears,
  antaresfbzone,
  thresholdFilter,
  core_ahc,
  log_detail
)
}
\arguments{
\item{sim_opts}{Simulation options, as returned by antaresRead::setSimulationPath}

\item{areas}{(string or vector of strings) what areas the patch should
be applied on.}

\item{virtual_areas}{(string or vector of strings) Virtual areas of the study,
excluded from the patch.}

\item{links_NTC_data}{NTC}

\item{ptdf_FB_data}{ptdf}

\item{capacity_FB_data}{capa}

\item{ts_FB_data}{ts}

\item{mcYears}{(numeric or vector of numeric) The Monte-Carlo years to
extract from. The special value "all" extracts all Monte-Carlo Years.}

\item{antaresfbzone}{name for new antares area}

\item{thresholdFilter}{filtering to important modification}

\item{core_ahc}{(boolean) TRUE for using CORE area, FALSE for using CWE area}

\item{log_detail}{(boolean) TRUE for more details}
}
\description{
Applies the Adequacy and write study by mc year
}
