% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{apply_adq_patch}
\alias{apply_adq_patch}
\title{Applies the Adequacy Patch on a given simulation}
\usage{
apply_adq_patch(
  sim_opts = antaresRead::simOptions(),
  areas = "all",
  virtual_areas = NULL,
  mcYears = "all",
  ptdf_FB_data = NULL,
  capacity_FB_data = NULL,
  ts_FB_data = NULL,
  core_ahc
)
}
\arguments{
\item{sim_opts}{(string) Simulation options, as returned by antaresRead::setSimulationPath}

\item{areas}{(string or vector of strings) what areas the patch should
be applied on. Default: ""}

\item{virtual_areas}{(string or vector of strings) Virtual areas of the study,
excluded from the patch. Default: NULL}

\item{mcYears}{(numeric or vector of numeric) The Monte-Carlo years to
extract from. The special value "all" extracts all Monte-Carlo Years.
Default: "all"}

\item{ptdf_FB_data}{ptdf_FB_data}

\item{capacity_FB_data}{capacity_FB_data}

\item{ts_FB_data}{ts_FB_data}
}
\value{
(data.table) Table giving the MRG, ENS and net-position for each country
at each time-step
}
\description{
The Adequacy patch is a post-processing phase on an Antares study simulation,
applying the local-matching and curtailment sharing rules as defined by the
EUPHEMIA to correct situations with at least one country in loss of load.
}
