% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adq_patch_core.R
\name{.single_time_step_core}
\alias{.single_time_step_core}
\title{Calls the AMPL Adequacy patch at each time-step}
\usage{
.single_time_step_core(ampl, patch, capacity)
}
\arguments{
\item{ampl}{(rAMPL::AMPL) AMPL object, with model and PTDF already loaded}

\item{patch}{(data.table) Relevant data form the simulation for this time-step}

\item{capacity}{(data.table) Limit capacity on each CB, containing both flow-based
and NTC data}
}
\value{
(data.table) Table giving the MRG, ENS and net-position for each country
at this time-step
}
\description{
Calls the AMPL Adequacy patch at each time-step
}
