% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_capa.R
\name{extract_FB_capacity}
\alias{extract_FB_capacity}
\title{Extracts the maximum transfer capacity on each CB for the flow-based domain of a study}
\usage{
extract_FB_capacity(sim_opts = antaresRead::simOptions())
}
\arguments{
\item{sim_opts}{(list) Simulation options as given by antaresRead::setSimulationPath}
}
\value{
(data.table) Table containing the limit capacity for each critical
branch on each typical day and for each hour.
}
\description{
Extracts the maximum transfer capacity on each CB for the flow-based domain of a study
}
\examples{
\dontrun{
sim_opts = antaresRead::setSimulationPath("path/to/my/simulation")

capacity_FB_data = extract_FB_capacity(sim_opts=sim_opts)
}
}
