% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_PTDF.R
\name{extract_FB_ptdf}
\alias{extract_FB_ptdf}
\title{Extracts the Power Transmission and Distribution Flows on each CB for each
country for the flow-based domain of a study.
It also converts the initial PTDFs, given for each link in PTDFs fiven for
each country.}
\usage{
extract_FB_ptdf(sim_opts = antaresRead::simOptions(), patch_alegro = T)
}
\arguments{
\item{sim_opts}{(list) Simulation options as given by antaresRead::setSimulationPath}
}
\value{
(data.table) Table containing the PTDFs of each country for each
critical branch
}
\description{
Extracts the Power Transmission and Distribution Flows on each CB for each
country for the flow-based domain of a study.
It also converts the initial PTDFs, given for each link in PTDFs fiven for
each country.
}
\examples{
\dontrun{
sim_opts = antaresRead::setSimulationPath("path/to/my/simulation")

ptdf_FB_data = extract_FB_ptdf(sim_opts=sim_opts)
}
}
