% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NTC.R
\name{extract_NTC_links}
\alias{extract_NTC_links}
\title{Extracts the NTC links data from a study and formats the like the flow-based data}
\usage{
extract_NTC_links(
  areas = NULL,
  sim_opts = antaresRead::simOptions(),
  set_capacity_one_to_infinite = T
)
}
\arguments{
\item{areas}{(string or vector of strings) Areas between which we want to
extract the links.}

\item{sim_opts}{(list) Simulation options as given by antaresRead::setSimulationPath}
}
\value{
(list) such that $capacity is a data.table containing the maximum
transfer capacity for each link (divided in Direct and Indirect)
and $ptdf is a data.table containing, for each link, a PTDF of 1 for the
origin country of the link if it is direct, or for the destination country
if it is indirect.
}
\description{
Extracts the NTC links data from a study and formats the like the flow-based data
}
\examples{
\dontrun{
sim_opts = antaresRead::setSimulationPath("path/to/my/simulation")
areas = antaresRead::getAreas()

links_NTC_data = extract_NTC_links(areas=areas, sim_opts=sim_opts)
}
}
