% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patch.R
\name{extract_patch}
\alias{extract_patch}
\title{Extracts the data relevant for the Adequacy patch for a simulation output.}
\usage{
extract_patch(
  areas,
  virtual_areas,
  mcYears = "all",
  sim_opts = antaresRead::simOptions()
)
}
\arguments{
\item{areas}{(string or vector of strings) what areas the patch should
be applied on. Default: ""}

\item{virtual_areas}{(string or vector of strings) Virtual areas of the study,
excluded from the patch. Default: NULL}

\item{mcYears}{(numeric or vector of numeric) The Monte-Carlo years to
extract from. The special value "all" extracts all Monte-Carlo Years.
Default: "all"}

\item{sim_opts}{(list) Simulation options as given by antaresRead::setSimulationPath}
}
\value{
(data.table) Table containing, for each mcYear, time-step and country,
the DENS (domestic Energy Not Served) and DMRG (Domestic Margin).
}
\description{
It selects the time-steps in an ANtares study when at least one country is in
loss of load.
}
\examples{
\dontrun{
sim_opts = antaresRead::setSimulationPath("path/to/my/simulation")
areas = antaresRead::getAreas()

patch_data = extract_patch(areas=areas, mcYears=c(1, 3), sim_opts=sim_opts)
}


}
