% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_adq}
\alias{run_adq}
\title{Applies the Adequacy Patch on a study}
\usage{
run_adq(
  opts,
  areas,
  virtual_areas,
  mcYears,
  ext = NULL,
  nbcl = 10,
  antaresfbzone = "model_description_fb",
  showProgress = TRUE,
  thresholdFilter = 1e+06,
  fb_constraints = TRUE
)
}
\arguments{
\item{opts}{Simulation options, as returned by antaresRead::setSimulationPath}

\item{areas}{(string or vector of strings) what areas the patch should
be applied on. Default: ""}

\item{virtual_areas}{(string or vector of strings) Virtual areas of the study,
excluded from the patch. Default: NULL}

\item{mcYears}{(numeric or vector of numeric) The Monte-Carlo years to
extract from. The special value "all" extracts all Monte-Carlo Years.
Default: "all"}

\item{ext}{name extand for output study.}

\item{nbcl}{numeric, number of process in cluster}

\item{antaresfbzone}{antares names of flowbased zone}

\item{showProgress}{show progress}

\item{thresholdFilter}{filtering to important modification}

\item{fb_constraints}{boolean: are there Flow-Based constraints to take into account in the Antares study}
}
\description{
Applies the Adequacy Patch on a study
}
\examples{
\dontrun{
opts <- setSimulationPath("path", 4)

areas = c("fr", "lu", "de", "cz", "pl", "ch", "at", "itn", "nl", "be", "es", "non", "se1", "model_description_fb_adq")
virtual_areas = getAreas(select = "_", regexpSelect = TRUE, exclude = c("model_description_fb", "x_open_turb", "x_open_pump"), regexpExclude = FALSE)
run_adq(opts, areas, virtual_areas, 1)
}

}
