% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_adq}
\alias{run_adq}
\title{Applies the Adequacy Patch on a study}
\usage{
run_adq(
  opts,
  areas,
  virtual_areas,
  mcYears,
  ext = NULL,
  nbcl = 10,
  antaresfbzone = "model_description_fb",
  showProgress = TRUE,
  thresholdFilter = 1e+06,
  core_ahc = FALSE,
  calculate_mc_all = FALSE,
  log_detail = FALSE
)
}
\arguments{
\item{opts}{Simulation options, as returned by antaresRead::setSimulationPath}

\item{areas}{(string or vector of strings) what areas the patch should
be applied on. Default: ""}

\item{virtual_areas}{(string or vector of strings) Virtual areas of the study,
excluded from the patch. Default: NULL}

\item{mcYears}{(numeric or vector of numeric) The Monte-Carlo years to
extract from. The special value "all" extracts all Monte-Carlo Years.
Default: "all"}

\item{ext}{name extand for output study. Default: NULL}

\item{nbcl}{numeric, number of process in cluster Default: 10}

\item{antaresfbzone}{antares names of flowbased zone Default: "model_description_fb"}

\item{showProgress}{show progress Default: TRUE}

\item{thresholdFilter}{filtering to important modification Default: 1000000}

\item{core_ahc}{(boolean) TRUE for using CORE area, FALSE for using CWE area Default: FALSE}

\item{calculate_mc_all}{(boolean) TRUE carries out the timestep aggregation of the output Default: FALSE}

\item{log_detail}{(boolean) TRUE for more details, Default: FALSE}
}
\description{
#' The Adequacy patch is a post-processing phase on an Antares study simulation,
applying the local-matching and curtailment sharing rules as defined by the
EUPHEMIA to correct situations with at least one country in loss of load.
}
\examples{
\dontrun{
opts <- setSimulationPath("path", 4)

areas = c("fr", "lu", "de", "cz", "pl", "ch", "at", "itn", "nl", "be", "es", "non", "se1", "model_description_fb_adq")
virtual_areas = getAreas(select = "_", regexpSelect = TRUE, exclude = c("model_description_fb", "x_open_turb", "x_open_pump"), regexpExclude = FALSE)
run_adq(opts, areas, virtual_areas, 1)
}

}
